refactor: created components/client folder
