--- src/components/UserDetails/index.ts
@@ -1,6 +1,0 @@
-export { default as hideDetails } from "@/components/UserDetails/utils/hideDetails";
-
-export { useDetailsForm } from "@/components/UserDetails/hooks/useDetailsForm";
-export { updateUserAction } from "@/components/UserDetails/actions/updateUserAction";
-
-export { default as Details } from "@/components/UserDetails/components/common/Details";
