+++ src/components/client/Auth/components/Credentials/CredentialsLogin.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { useLogin } from "@/components/Auth";
-import CredentialsForm from "@/components/Auth/components/Credentials/CredentialsForm";
+import { useLogin } from "@/components/client/Auth";
+import CredentialsForm from "@/components/client/Auth/components/Credentials/CredentialsForm";
 
 export function CredentialsLogin() {
   const { error, handleLogin } = useLogin();
