+++ src/components/client/Auth/components/OAuth/OAuthLogin.tsx
@@ -1,5 +1,5 @@
-import { AvailableOAuthProviders } from "@/components/Auth";
-import OAuthForm from "@/components/Auth/components/OAuth/OAuthForm";
+import { AvailableOAuthProviders } from "@/components/client/Auth";
+import OAuthForm from "@/components/client/Auth/components/OAuth/OAuthForm";
 
 export function OAuthLogin() {
   return (
