+++ src/components/client/UserDetails/components/DetailsPhone.tsx
@@ -1,6 +1,10 @@
 "use client";
 
-import { Details, hideDetails, useDetailsForm } from "@/components/UserDetails";
+import {
+  Details,
+  hideDetails,
+  useDetailsForm,
+} from "@/components/client/UserDetails";
 
 type DetailsPhoneProps = {
   number: string;
