+++ src/components/client/UserDetails/components/common/DetailsOverlay.tsx
@@ -1,5 +1,7 @@
 import OverlayPortal from "@/components/common/Overlay";
 import { Dispatch, ReactNode, SetStateAction } from "react";
+
+// TODO: ICON
 import { CloseIcon } from "next/dist/client/components/react-dev-overlay/internal/icons/CloseIcon";
 
 type DetailsOverlayProps = {
@@ -18,7 +20,6 @@ export function DetailsOverlay({
   return (
     <OverlayPortal
       isOpen={isEditing}
-      setIsOpen={setIsEditing}
       className="z-50 bg-boldColor bg-opacity-50"
     >
       <div className="absolute left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 rounded-md bg-white p-8 pt-4 shadow-2xl">
