+++ src/components/client/UserDetails/components/common/DetailsView.tsx
