+++ src/components/client/UserDetails/index.ts
@@ -1,0 +1,6 @@
+export { default as hideDetails } from "@/components/client/UserDetails/utils/hideDetails";
+
+export { useDetailsForm } from "@/components/client/UserDetails/hooks/useDetailsForm";
+export { updateUserAction } from "@/components/client/UserDetails/actions/updateUserAction";
+
+export { default as Details } from "@/components/client/UserDetails/components/common/Details";
