+++ src/components/client/UserDetails/utils/hideDetails.ts
